refactor: extract session types, errors, builder into session/types.rs
Move ~375 lines of pure type definitions from session/mod.rs:

- SessionConfig, FrameProgress, StepResult, SessionMetrics, ErrorPolicy
- SessionError enum (with all variants)
- DetectionSink trait + blanket impl
- StitchSessionBuilder struct + impl
- compute_frame_limit helper

session/mod.rs: 1,950 -> 1,575 lines. StitchSession struct and all
its methods remain in mod.rs. Re-exports preserve external API.

14 tests pass, clippy clean.
